fix(reg): [iOS] Fix pointer event was raised on SCP parent before SCP children
