+++ src/Uno.UI/UI/Xaml/Controls/ScrollContentPresenter/NativeScrollContentPresenter.Xamarin.cs
@@ -25,7 +25,7 @@ using View = Windows.UI.Xaml.UIElement;
 
 namespace Windows.UI.Xaml.Controls
 {
-	internal sealed partial class NativeScrollContentPresenter : IScrollContentPresenter
+	internal partial class NativeScrollContentPresenter : IScrollContentPresenter
 	{
 		private View _content;
