+++ src/Uno.UI/UI/Xaml/Controls/ScrollContentPresenter/NativeScrollContentPresenter.iOS.cs
@@ -15,7 +15,9 @@ using UIKit;
 using CoreGraphics;
 using Uno.Logging;
 using Windows.Foundation;
+using Windows.UI.Input;
 using Uno.UI.Xaml.Input;
+using DraggingEventArgs = UIKit.DraggingEventArgs;
 
 namespace Windows.UI.Xaml.Controls
 {
@@ -51,9 +53,6 @@ namespace Windows.UI.Xaml.Controls
 			{
 				GetParentScrollViewer()?.UpdateDimensionProperties();
 			};
-
-			// The UIScrollView mutes all touch events on super views ...
-			AddGestureRecognizer(new PassThroughPseudoGestureRecognizer(this));
 		}
 
 		public NativeScrollContentPresenter()
@@ -506,6 +505,48 @@ namespace Windows.UI.Xaml.Controls
 
 		#region Touches
 
+		private UIElement _touchTarget;
+
+		/// <inheritdoc />
+		public override void TouchesBegan(NSSet touches, UIEvent evt)
+		{
+			base.TouchesBegan(touches, evt);
+
+			// We wait for the first touches to get the parent so we don't have to track Loaded/UnLoaded
+			// Like native dispatch on iOS, we do "implicit captures" the target.
+			if (this.GetParent() is UIElement parent)
+			{
+				_touchTarget = parent;
+				_touchTarget.TouchesBegan(touches, evt);
+			}
+		}
+
+		/// <inheritdoc />
+		public override void TouchesMoved(NSSet touches, UIEvent evt)
+		{
+			base.TouchesMoved(touches, evt);
+
+			_touchTarget?.TouchesMoved(touches, evt);
+		}
+
+		/// <inheritdoc />
+		public override void TouchesEnded(NSSet touches, UIEvent evt)
+		{
+			base.TouchesEnded(touches, evt);
+
+			_touchTarget?.TouchesEnded(touches, evt);
+			_touchTarget = null;
+		}
+
+		/// <inheritdoc />
+		public override void TouchesCancelled(NSSet touches, UIEvent evt)
+		{
+			base.TouchesCancelled(touches, evt);
+
+			_touchTarget?.TouchesCancelled(touches, evt);
+			_touchTarget = null;
+		}
+
 		/*
 		 * By default the UIScrollView will delay the touches to the content until it detects
 		 * if the manipulation is a drag. And even there, if it detects that the manipulation
@@ -544,6 +585,12 @@ namespace Windows.UI.Xaml.Controls
 			{
 				_scrollPresenter = scrollPresenter;
 			}
+
+			/// <inheritdoc />
+			protected override bool CanConflict(GestureRecognizer.Manipulation manipulation)
+				=> _scrollPresenter.CanHorizontallyScroll && manipulation.IsTranslateXEnabled
+				|| _scrollPresenter.CanVerticallyScroll && manipulation.IsTranslateYEnabled
+				|| manipulation.IsDragManipulation; // This will actually always be false when CanConflict is being invoked in current setup.
 
 			/// <inheritdoc />
 			protected override void SetCanDelay(bool canDelay)
